NVPL LAPACK Solver Support on ARM
* NVPL LAPACK support for all decompositions on ARM

* Simplify SVD interface to hide column-major logic

* More comprehensive type and shape checks for all decompositions

* Batched tests and complex tests for remaining decompositions

* Fixed sign bug in determinant

* Split solver into separate files

+++ test/test_vectors/generators/matx_common.py
@@ -18,6 +18,6 @@ def to_file(var, name):
 
 def randn_ndarray(tshape, dtype):
     if np.issubdtype(dtype, np.floating):
-        return np.random.randn(*tshape)
+        return np.random.randn(*tshape).astype(dtype)
     else:
-        return np.random.randn(*tshape) + 1j*np.random.randn(*tshape)
+        return (np.random.randn(*tshape) + 1j*np.random.randn(*tshape)).astype(dtype)
